Do not overwrite existing backup
This checks the existence of the backup file to avoid overwriting any
existing files.

A backup file can already exist when two or more backups are run within
the same second (e.g. when using lineinfile and with_items).

+++ lib/ansible/module_utils/basic.py
@@ -1596,13 +1596,12 @@ class AnsibleModule(object):
             # backups named basename-YYYY-MM-DD@HH:MM:SS~
             ext = time.strftime("%Y-%m-%d@%H:%M:%S~", time.localtime(time.time()))
             backupdest = '%s.%s' % (fn, ext)
-
-            try:
-                shutil.copy2(fn, backupdest)
-            except (shutil.Error, IOError):
-                e = get_exception()
-                self.fail_json(msg='Could not make backup of %s to %s: %s' % (fn, backupdest, e))
-
+            if not os.path.exists(backupdest):
+                try:
+                    shutil.copy2(fn, backupdest)
+                except (shutil.Error, IOError):
+                    e = get_exception()
+                    self.fail_json(msg='Could not make backup of %s to %s: %s' % (fn, backupdest, e))
         return backupdest
 
     def cleanup(self, tmpfile):
